Add Files
This displays a gallery of random photos from a NASA database. When you press one, it expands the image, and when you click one more time, it goes back. You can also log in to a new gallery at any time by pressing the button at the bottom of the canvas.